requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
praw==7.7.1
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
wordcloud==1.9.2
textblob==0.17.1
vaderSentiment==3.3.2
groq==0.4.2
google-generativeai==0.3.2
python-multipart==0.0.6
reportlab==4.0.7
weasyprint==60.2
Pillow==10.1.0 